backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.2
psycopg2-binary==2.9.10
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
